refactor to consolidate gs uri parsing logic
