+++ src/main/java/bio/terra/service/dataset/DataDeletionRequestValidator.java
@@ -3,8 +3,8 @@ package bio.terra.service.dataset;
 import bio.terra.model.DataDeletionGcsFileModel;
 import bio.terra.model.DataDeletionRequest;
 import bio.terra.model.DataDeletionTableModel;
+import bio.terra.service.common.gcs.GcsUriUtils;
 import bio.terra.service.dataset.exception.InvalidUriException;
-import bio.terra.service.dataset.flight.ingest.IngestUtils;
 import javax.validation.constraints.NotNull;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
@@ -30,7 +30,7 @@ public class DataDeletionRequestValidator implements Validator {
       errors.rejectValue("tables.gcsFileSpec", "FileSpecMissing", "Requires a file spec");
     } else {
       try {
-        IngestUtils.parseBlobUri(gcsFileSpec.getPath());
+        GcsUriUtils.parseBlobUri(gcsFileSpec.getPath());
       } catch (InvalidUriException ex) {
         errors.rejectValue("tables.gcsFileSpec.path", "InvalidGsUri", ex.getMessage());
       }
